+++ tests/Notifications/LineChannelTest.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Tests\Notifications;
+
+use Illuminate\Notifications\AnonymousNotifiable;
+use Mockery;
+use Revolution\Line\Facades\Bot;
+use Revolution\Line\Notifications\LineChannel;
+use Revolution\Line\Notifications\LineMessage;
+use Tests\Notifications\Fixtures\LineNotificationStub;
+use Tests\Notifications\Fixtures\LineNotifyStub;
+use Tests\Notifications\Fixtures\TestNotifiableStub;
+use Tests\TestCase;
+
+class LineChannelTest extends TestCase
+{
+    public function testLineChannel()
+    {
+        Bot::shouldReceive('pushMessage')->once();
+
+        $channel = new LineChannel();
+
+        $notifiable = (new AnonymousNotifiable())
+            ->route('line', 'test');
+
+        $channel->send($notifiable, new LineNotificationStub('test'));
+        $channel->send(new TestNotifiableStub(), new LineNotificationStub('test'));
+    }
+}